ESP32-S3 support, Fire Tier 1 improvements, docs/tooling updates
ESP32-S3 platform:
- Add XIAO_ESP32S3 target to Makefile and BlinkyImplementations.h
- Add ESP32-S3 HAL implementations (DefaultHal.h, ISystemTime.h)
- Use raw ESP-IDF I2S PDM-RX driver with software gain in Esp32PdmMic

Fire generator Tier 1:
- Phase-driven thermal buoyancy breathing (0.5 off-beat → 1.0 on-beat)
- Downbeat transient effects: spread mult, color shift, velocity boost with decay
- Beat-in-measure spawn bias (left/right alternation)
- onsetDensity-driven lifespan, noise speed, wind turbulence

Audio / mic:
- Nrf52PdmMic: expose poll() as no-op (base class default)
- AudioController: minor header updates

Build tooling:
- uf2_upload.py: improved UF2 upload script for ESP32-S3
- Makefile: add flash targets for both platforms

Docs:
- IMPROVEMENT_PLAN.md: add ESP32-S3 mic calibration priority
- BUILD_GUIDE.md, DEVELOPMENT.md: ESP32-S3 build/flash procedures

+++ blinky-things/blinky-things.ino
@@ -214,7 +214,7 @@ void setup() {
     }
 
     leds->begin();
-    leds->setBrightness(min(config.matrix.brightness, 255));
+    leds->setBrightness(min((int)config.matrix.brightness, 255));
     leds->show();
 
     // Basic LED test - light up first few LEDs to verify hardware
